requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
cloudpickle==2.2.0
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.33
fonttools==4.38.0
gym==0.26.2
gym-notices==0.0.8
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.3
numpy==1.24.1
packaging==23.0
Pillow==9.4.0
pybind11==2.10.3
pygame==2.1.2
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.2
setuptools-scm==7.1.0
six==1.16.0
tomli==2.0.1
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14